# Conflicts:
#	libarchive/archive_read_support_format_rar.c
#	libarchive/archive_read_support_format_tar.c
#	libarchive/test/test_ar_mode.c
#	test_utils/test_common.h
#	test_utils/test_main.c

+++ libarchive/archive_acl.c
@@ -990,7 +990,7 @@ archive_acl_to_text_l(struct archive_acl *acl, ssize_t *text_len, int flags,
 
 	len = strlen(s);
 
-	if ((ssize_t)len > (length - 1))
+	if (len > length - 1)
 		__archive_errx(1, "Buffer overrun");
 
 	if (text_len != NULL)
